+++ src/context/cartItems/CartItemsContext.ts
@@ -1,0 +1,11 @@
+import { createContext } from 'react';
+import { CartItem } from '../../types';
+
+interface CartItemsContextValue {
+  cartItems: CartItem[];
+  refreshCartItems: () => void;
+}
+
+export const CartItemsContext = createContext<CartItemsContextValue | null>(
+  null
+);
